delete attack for comparing bloom filter to broadcasted transaction; fold into II.A
Source: OBPP Network Attack Notes google doc, non-reorganized section.
The Note in this doc says:
> Notes: Daniel suggests removing this one, see GitHub issue
https://github.com/OpenBitcoinPrivacyProject/wallet-ratings/issues/135
> TODO: This looks like a good idea, already captured in criterion for
II.A. Need to move CM23 to that attack or elsewhere.

I moved CM23 to attack II.A and added an example scenario to help
indicate that this scenario falls under that attack now.

Note that this attack was excluded from the reorganized section of the
doc on the basis of deleting it.