List CM32 under attacks II.B, II.D, II.E, and II.I. Also, add CR162
Source: OBPP Network Attack Notes

> Similar to CR160, add new criteria for server authenticating to
client with a CA-issued certificate. Less effective than pinning.
Remove this attack, and list CM32 and its current criteria under
reorganized attacks B, D, E, and I. Note: CR161 might describe the
behavior of BIP 151 (if so note this in a JSON comment. if not, maybe
need to rephrase it, e.g. “The wallet client uses a symmetric key to
authenticate to the server that was pre-shared or securely exchanged
via a secure key exchange protocol”). This countermeasure doesn’t
necessarily protect user against all network observers, but may limit
to those the client is directly connected and thwart passive observers.