fold attack III.B into II.E
1. II.E description is generalized to describe an action rather than
the focus on linking addresses specifically.
2. Add new countermeasure for using a mixing protocol that can’t be
fingerprinted and list that under II.E
3. Move all countermeasures/criteria from III.B to II.E and delete
III.B.

This is a change that Daniel, Craig, and I agreed in our last meeting
on the basis that, since attack III.B described an action that was the
essential component of the action described in II.E, it would be
cleaner to fold it in.

Now it’s clear that, once you perform this action, you may also
compromise the fingerprint information.

We may need to copy the attack to wallet provider since some
countermeasures here do not work when the network observer is the
wallet provider e.g. for a web wallet.

We’ve also discussed folding attack III.A into a network observer
attack and may do that next.

syntax correctness:
```
$ python validate_json.py obpp3.hjson
'obpp3.hjson' matches the specified schema.
```